requirements.txt
openai>=1.40.0
python-dotenv>=1.0
pydantic>=2.6
pyyaml>=6
rich>=13
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
pyjwt>=2.9.0
cryptography>=43.0.1
requests>=2.31
pandas>=2.2
pymupdf>=1.24.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.12
python-jose[cryptography]>=3.3
authlib>=1.2
httpx>=0.24